chore: refactor dialog
