+++ packages/grid/src/elements/simple-html-grid-panel.ts
@@ -2,7 +2,7 @@ import { customElement } from '@simple-html/core';
 import { GridInterface } from '../gridInterface';
 import { html, svg } from 'lit-html';
 import { SimpleHtmlGrid } from '..';
-import { generateMenu } from './generateMenu';
+import { generateMenuWithComponentName } from './generateMenuWithComponentName';
 
 @customElement('simple-html-grid-panel')
 export default class extends HTMLElement {
@@ -23,7 +23,12 @@ export default class extends HTMLElement {
         }
         if (e.type === 'contextmenu') {
             e.preventDefault();
-            generateMenu('simple-html-grid-menu-panel', e, this.connector, this.ref);
+            generateMenuWithComponentName(
+                'simple-html-grid-menu-panel',
+                e,
+                this.connector,
+                this.ref
+            );
         }
     }
